Extract `LinkOption` component
